+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/broker/AbstractAdvancedBrokerTest.java
@@ -4,9 +4,12 @@ import org.junit.Test;
 import org.keycloak.admin.client.resource.IdentityProviderResource;
 import org.keycloak.admin.client.resource.RealmResource;
 import org.keycloak.admin.client.resource.UserResource;
+import org.keycloak.common.Profile;
 import org.keycloak.common.util.Time;
+import org.keycloak.models.AuthenticationExecutionModel;
 import org.keycloak.models.IdentityProviderMapperSyncMode;
 import org.keycloak.models.IdentityProviderSyncMode;
+import org.keycloak.representations.idm.AuthenticationExecutionInfoRepresentation;
 import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.idm.ComponentRepresentation;
 import org.keycloak.representations.idm.IdentityProviderMapperRepresentation;
@@ -17,6 +20,7 @@ import org.keycloak.representations.idm.UserRepresentation;
 import org.keycloak.services.Urls;
 import org.keycloak.storage.UserStorageProvider;
 import org.keycloak.testsuite.Assert;
+import org.keycloak.testsuite.ProfileAssume;
 import org.keycloak.testsuite.admin.ApiUtil;
 import org.keycloak.testsuite.federation.DummyUserFederationProviderFactory;
 import org.keycloak.testsuite.util.ClientBuilder;
@@ -32,6 +36,7 @@ import javax.ws.rs.core.Response;
 import java.net.URI;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
@@ -65,6 +70,8 @@ import static org.keycloak.testsuite.broker.BrokerTestTools.waitForPage;
  */
 public abstract class AbstractAdvancedBrokerTest extends AbstractBrokerTest {
 
+    private static final String Browser_Conditional_OTP = "Browser - Conditional OTP";
+    private static final String Browser = "browser";
 
     protected void createRoleMappersForConsumerRealm() {
         createRoleMappersForConsumerRealm(IdentityProviderMapperSyncMode.FORCE);
@@ -530,6 +537,8 @@ public abstract class AbstractAdvancedBrokerTest extends AbstractBrokerTest {
             realm.attackDetection().clearBruteForceForUser(userId);
 
             loginTotpPage.login(totp.generateTOTP(totpSecret));
+            //return to client flow -> otp has been configured -> ask again
+            loginTotpPage.login(totp.generateTOTP(totpSecret));
             waitForAccountManagementTitle();
             logoutFromRealm(getConsumerRoot(), bc.consumerRealmName());
         } finally {
@@ -576,6 +585,17 @@ public abstract class AbstractAdvancedBrokerTest extends AbstractBrokerTest {
      */
     @Test
     public void testWithLinkedFederationProvider() {
+        // don't run this test when map storage is enabled, as map storage doesn't support the legacy style federation
+        ProfileAssume.assumeFeatureDisabled(Profile.Feature.MAP_STORAGE);
+
+        //disable otp for browser flow
+        List<AuthenticationExecutionInfoRepresentation> executionReps = adminClient.realm(bc.consumerRealmName()).flows().getExecutions(Browser);
+        AuthenticationExecutionInfoRepresentation exec =  executionReps.stream().filter(authExec -> authExec.getDisplayName().equals(Browser_Conditional_OTP)).findFirst().orElse(null);
+        if (exec != null) {
+            exec.setRequirement(AuthenticationExecutionModel.Requirement.DISABLED.name());
+            adminClient.realm(bc.consumerRealmName()).flows().updateExecutions(Browser, exec);
+        }
+
         try {
             updateExecutions(AbstractBrokerTest::disableUpdateProfileOnFirstLogin);
 
@@ -616,6 +636,11 @@ public abstract class AbstractAdvancedBrokerTest extends AbstractBrokerTest {
         } finally {
             removeUserByUsername(adminClient.realm(bc.consumerRealmName()), "test-user");
             removeUserByUsername(adminClient.realm(bc.consumerRealmName()), "test-user-noemail");
+
+            if (exec != null) {
+                exec.setRequirement(AuthenticationExecutionModel.Requirement.CONDITIONAL.name());
+                adminClient.realm(bc.consumerRealmName()).flows().updateExecutions(Browser, exec);
+            }
         }
     }
 }